--- packages/client/src/pages/Forum/ForumTopic/ForumTopic.tsx
@@ -1,30 +1,0 @@
-import { FC } from 'react'
-import { Paper } from '@mui/material'
-import { forumTopicCommentMock, forumTopicMock } from './mock'
-import { TopicTitles } from '../components/TopicTitles'
-import { ForumTopicComment } from './ForumTopicComment/ForumTopicComment'
-import './style.scss'
-
-export const ForumTopic: FC = () => {
-  /// тут будет useParams для получения всего что нужно по айди
-  return (
-    <Paper elevation={5}>
-      <TopicTitles
-        title={forumTopicMock.title}
-        text={forumTopicMock.text}
-        avatar={forumTopicMock.avatar}
-        userName={forumTopicMock.userName}
-      />
-      <div className="forum-topic-comments">
-        {forumTopicCommentMock.map((comment, index) => (
-          <ForumTopicComment
-            avatar={comment.avatar}
-            userName={comment.userName}
-            text={comment.text}
-            key={comment.userName + index}
-          />
-        ))}
-      </div>
-    </Paper>
-  )
-}
